move elm-readers;py 3 experiments;
