+++ elm/model_selection/__init__.py
@@ -1,6 +1,5 @@
 from elm.model_selection.sklearn_support import (PARTIAL_FIT_MODEL_DICT,
                                                  PARTIAL_FIT_MODEL_STR,
                                                  MODELS_WITH_PREDICT_DICT,
-                                                 UNSUPERVISED_MODEL_STR,
                                                  DECOMP_MODEL_STR)
-from elm.model_selection.util import get_args_kwargs_defaults
+from elm.config.func_signatures import get_args_kwargs_defaults
