+++ elm/pipeline/predict_many.py
@@ -11,9 +11,10 @@ import xarray as xr
 
 
 from elm.config import import_callable, parse_env_vars
-from elm.readers import inverse_flatten, ElmStore
+from elm.sample_util import inverse_flatten, ElmStore
 from elm.sample_util.samplers import make_samples_dask
 from elm.pipeline.util import _next_name
+from elm.sample_util.elm_store import check_X_data_type
 
 logger = logging.getLogger(__name__)
 
@@ -27,8 +28,7 @@ def _predict_one_sample_one_arg(estimator,
                                 elm_predict_path,
                                 X_y_sample_weight):
     X, y, sample_weight = X_y_sample_weight
-    if not isinstance(X, (ElmStore, xr.Dataset)):
-        raise ValueError('Expected an ElmStore or xarray.Dataset')
+    check_X_data_type(X)
     out = []
     prediction, X_final = estimator.predict(X, return_X=True)
     if prediction.ndim == 1:
