--- elm/readers/__init__.py
@@ -1,11 +1,0 @@
-'''Package of readers from common satellite and weather data formats'''
-# The modules below use __all__
-from elm.readers.elm_store import *
-from elm.readers.hdf4 import *
-from elm.readers.hdf5 import *
-from elm.readers.netcdf import *
-from elm.readers.tif import *
-from elm.readers.util import *
-from elm.readers.reshape import *
-from elm.readers.load_array import *
-from elm.readers.local_file_iterators import *
