+++ elm/sample_util/change_coords.py
@@ -9,15 +9,15 @@ import numpy as np
 
 from elm.sample_util.step_mixin import StepMixin
 from elm.config import ElmConfigError, import_callable
-from elm.readers import (select_canvas as _select_canvas,
-                         drop_na_rows as _drop_na_rows,
-                         ElmStore,
-                         flatten as _flatten,
-                         inverse_flatten as _inverse_flatten,
-                         Canvas,
-                         check_is_flat,
-                         transpose as _transpose,
-                         aggregate_simple)
+from elm.sample_util.reshape import (select_canvas as _select_canvas,
+                                     drop_na_rows as _drop_na_rows,
+                                     ElmStore,
+                                     flatten as _flatten,
+                                     inverse_flatten as _inverse_flatten,
+                                     Canvas,
+                                     check_is_flat,
+                                     transpose as _transpose,
+                                     aggregate_simple)
 
 CHANGE_COORDS_ACTIONS = (
     'select_canvas',
