+++ elm/sample_util/preproc_scale.py
@@ -14,8 +14,8 @@ import numpy as np
 import xarray as xr
 
 from elm.config import import_callable
-from elm.model_selection.util import get_args_kwargs_defaults
-from elm.readers import *
+from elm.config.func_signatures import get_args_kwargs_defaults
+from elm.sample_util import ElmStore
 from elm.sample_util.step_mixin import StepMixin
 
 class SklearnBase(StepMixin):
