+++ elm/sample_util/tests/test_change_coords.py
@@ -4,10 +4,11 @@ import xarray as xr
 from sklearn.decomposition import PCA
 
 from elm.config import ConfigParser
-from elm.pipeline.tests.util import (random_elm_store,
-                                     test_one_config as tst_one_config,
+from elm.pipeline.tests.util import (test_one_config as tst_one_config,
                                      tmp_dirs_context)
-from elm.readers import *
+from elm.sample_util.make_blobs import random_elm_store
+from elm.sample_util.reshape import *
+from elm.sample_util.elm_store import ElmStore
 from elm.pipeline import Pipeline
 
 X = random_elm_store()
