+++ elm/sample_util/tests/test_ts_grid_tools.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 import xarray as xr
 
-from elm.readers import *
+from elm.sample_util.elm_store import ElmStore
 from elm.pipeline import steps
 
 def make_3d():
