+++ elm/sample_util/transform.py
@@ -12,7 +12,7 @@ import numpy as np
 import xarray as xr
 
 from elm.sample_util.step_mixin import StepMixin
-from elm.readers import ElmStore
+from elm.sample_util.elm_store import ElmStore
 
 logger = logging.getLogger(__name__)
